(Breaking change) Sprite Props (#19)
This updates all packages and docs to allow all common props like scaleX to be used on all Sprites, not just Textures. This also removes the outer nesting of the position prop.

anchorX and anchorY also updated to be absolute, not relative. Adds an align prop to text Textures to accommodate this.