Native Sprites & TextInput (#27)
This adds Native Sprites and an implementation of one for text input fields.

Native Sprites allow anyone to easily build features into Replay it doesn't support. In theory it could be used for anything supported in the browser or mobile device (like cameras, ML tools or even a 3D renderer).

Some bug fixes in the Swift package are also included in this commit.